+++ test/crypto_sign/checksums.c
@@ -27,11 +27,11 @@ typedef struct state {
   uint8_t *m2;
   uint8_t *c2;
   uint8_t *t2;
-  uint64_t plen;
-  uint64_t slen;
-  uint64_t mlen;
-  uint64_t clen;
-  uint64_t tlen;
+  size_t plen;
+  size_t slen;
+  size_t mlen;
+  size_t clen;
+  size_t tlen;
   void* free[10];
 } state;
 
@@ -66,7 +66,7 @@ state* preallocate(void)
 
 void allocate(state *s)
 {
-  uint64_t alloclen = 0;
+  size_t alloclen = 0;
   if (alloclen < TUNE_BYTES) alloclen = TUNE_BYTES;
   if (alloclen < MAXTEST_BYTES + JADE_SIGN_BYTES) alloclen = MAXTEST_BYTES + JADE_SIGN_BYTES;
   if (alloclen < JADE_SIGN_PUBLICKEYBYTES) alloclen = JADE_SIGN_PUBLICKEYBYTES;
